Handle conditional import for ixmp4 IO functions
